requirements.txt
django
psycopg2-binary
django-cors-headers
asyncpg
django-crispy-forms
twilio
websockets
websocket-client
pandas
djangorestframework
ib_insync
pandas-ta